create a fetch data function in a seperate file to combine the fetch logic in the open-source page
