+++ website/src/app/[lang]/[region]/(website)/open-source/(components)/get-commits.ts
@@ -1,3 +1,5 @@
+import { fetchData } from './fetch-data';
+
 const owner = 'socialincome-san';
 const repo = 'public';
 
@@ -23,50 +25,12 @@ export async function getCommits() {
 
 	// Fetch recent commits from the last 30 days
 	const commitUrl = `https://api.github.com/repos/${owner}/${repo}/commits?since=${startDateISO}&until=${endDate}`;
-	const headers: Record<string, string> = {
-		Accept: 'application/vnd.github+json',
-	};
-	// Conditionally add the Authorization header if GITHUB_PAT is available
-	if (process.env.GITHUB_PAT) {
-		headers['Authorization'] = `Bearer ${process.env.GITHUB_PAT}`;
-	}
-	const res = await fetch(commitUrl, {
-		headers,
-	});
-
-	if (!res.ok) {
-		const errorDetails = await res.text();
-		const status = res.status;
-		if (status === 403) {
-			throw new Error(
-				'GitHub API rate limit exceeded. Please try again later or increase rate limit by authenticating.',
-			);
-		} else if (status === 404) {
-			throw new Error(`GitHub repository ${owner}/${repo} not found.`);
-		} else {
-			throw new Error(`Failed to fetch recent commits from GitHub: ${status} - ${errorDetails}`);
-		}
-	}
-
-	const recentCommits: GitHubCommit[] = await res.json();
+	const recentCommitsRes = await fetchData(owner, repo, commitUrl);
+	const recentCommits: GitHubCommit[] = await recentCommitsRes.json();
 
 	// Fetch total commit count
 	const totalCommitsUrl = `https://api.github.com/repos/${owner}/${repo}/commits?per_page=1`;
-	const totalCommitsRes = await fetch(totalCommitsUrl, {
-		headers,
-	});
-
-	if (!totalCommitsRes.ok) {
-		const errorDetails = await totalCommitsRes.text();
-		const status = totalCommitsRes.status;
-		if (status === 403) {
-			throw new Error(`GitHub API rate limit exceeded: ${status} - ${errorDetails}.`);
-		} else if (status === 404) {
-			throw new Error(`GitHub repository ${owner}/${repo} not found while fetching total commits.`);
-		} else {
-			throw new Error(`Failed to fetch total commits from GitHub: ${status} - ${errorDetails}`);
-		}
-	}
+	const totalCommitsRes = await fetchData(owner, repo, totalCommitsUrl);
 
 	// Extract the last page number from the Link header to get the total commit count
 	const linkHeader = totalCommitsRes.headers.get('link');
