+++ website/src/app/[lang]/[region]/(website)/open-source/(components)/get-contributors.ts
@@ -1,3 +1,5 @@
+import { fetchData } from './fetch-data';
+
 const owner = 'socialincome-san';
 const repo = 'public';
 
@@ -17,31 +19,10 @@ interface GitHubContributor {
 	total: number;
 }
 
-export async function getContributors(): Promise<Contributor[]> {
-	const headers: Record<string, string> = {
-		Accept: 'application/vnd.github+json',
-	};
-	// Conditionally add the Authorization header if GITHUB_PAT is available
-	if (process.env.GITHUB_PAT) {
-		headers['Authorization'] = `Bearer ${process.env.GITHUB_PAT}`;
-	}
-	const res = await fetch(`https://api.github.com/repos/${owner}/${repo}/stats/contributors`, {
-		headers,
-	});
-
-	if (!res.ok) {
-		const errorDetails = await res.text();
-		const status = res.status;
-		if (status === 403) {
-			throw new Error(`GitHub API rate limit exceeded: ${status} - ${errorDetails}.`);
-		} else if (status === 404) {
-			throw new Error(`GitHub repository ${owner}/${repo} not found.`);
-		} else {
-			throw new Error(`Failed to fetch contributors from GitHub: ${status} - ${errorDetails}`);
-		}
-	}
-
-	const contributors = await res.json();
+export async function getContributors() {
+	const url = `https://api.github.com/repos/${owner}/${repo}/stats/contributors`;
+	const contributorsRes = await fetchData(owner, repo, url);
+	const contributors = await contributorsRes.json();
 
 	if (Object.keys(contributors).length === 0) {
 		console.warn('No contributor data available. The API returned an empty object.');
