+++ website/src/app/[lang]/[region]/(website)/open-source/(components)/get-forks.ts
@@ -1,3 +1,5 @@
+import { fetchData } from './fetch-data';
+
 const owner = 'socialincome-san';
 const repo = 'public';
 
@@ -8,28 +10,8 @@ interface GitHubFork {
 
 export async function getForkCount(): Promise<{ totalForks: number; newForks: number }> {
 	const repoUrl = `https://api.github.com/repos/${owner}/${repo}`;
-	const headers: Record<string, string> = {
-		Accept: 'application/vnd.github+json',
-	};
-	// Conditionally add the Authorization header if GITHUB_PAT is available
-	if (process.env.GITHUB_PAT) {
-		headers['Authorization'] = `Bearer ${process.env.GITHUB_PAT}`;
-	}
-	const repoRes = await fetch(repoUrl, { headers });
-
-	if (!repoRes.ok) {
-		const errorDetails = await repoRes.text();
-		const status = repoRes.status;
-		if (status === 403) {
-			throw new Error(`GitHub API rate limit exceeded: ${status} - ${errorDetails}.`);
-		} else if (status === 404) {
-			throw new Error(`GitHub repository ${owner}/${repo} not found.`);
-		} else {
-			throw new Error(`Failed to fetch repository info from GitHub: ${status} - ${errorDetails}`);
-		}
-	}
-
-	const repoData = await repoRes.json();
+	const repoDataRes = await fetchData(owner, repo, repoUrl);
+	const repoData = await repoDataRes.json();
 	const totalForks = repoData.forks_count;
 
 	// Calculate the date 30 days ago from today
@@ -44,25 +26,7 @@ export async function getForkCount(): Promise<{ totalForks: number; newForks: nu
 	let hasMore = true;
 
 	while (hasMore) {
-		const pagedRes = await fetch(`${forksUrl}&page=${page}`, {
-			headers,
-		});
-
-		if (!pagedRes.ok) {
-			const errorDetails = await pagedRes.text();
-			const status = pagedRes.status;
-
-			if (status === 403 && errorDetails.includes('API rate limit exceeded')) {
-				throw new Error(
-					'GitHub API rate limit exceeded during forks fetching. Please try again later or increase rate limit by authenticating.',
-				);
-			} else if (status === 404) {
-				throw new Error(`GitHub repository ${owner}/${repo} forks not found.`);
-			} else {
-				throw new Error(`Failed to fetch forks from GitHub: ${status} - ${errorDetails}`);
-			}
-		}
-
+		const pagedRes = await fetchData(owner, repo, `${forksUrl}&page=${page}`);
 		const forks: GitHubFork[] = await pagedRes.json();
 
 		// Count new forks within the last 30 days
